+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## [Unreleased] - ReleaseDate
+### Added
+- Added `BufRef::len`
+
 ### Changed
--  If an `lio_listio` operation fails asynchronously, the future will now
-   include final error status for all failed operations.
+- If an `lio_listio` operation fails asynchronously, the future will now
+  include final error status for all failed operations.
 
 ## [0.3.1] - 2018-07-01
 ### Fixed